@@ -2,7 +2,7 @@ package de.stephanlindauer.criticalmaps.handler;
 
 import android.app.Activity;
 import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import de.stephanlindauer.criticalmaps.R;
 import de.stephanlindauer.criticalmaps.service.ServerSyncService;
 
