@@ -1,7 +1,7 @@
 package de.stephanlindauer.criticalmaps.model;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 public class PermissionRequest {
     private final String permission;
