@@ -2,9 +2,8 @@ package de.stephanlindauer.criticalmaps.service;
 
 import android.app.Service;
 import android.content.Intent;
-import android.os.Build;
 import android.os.IBinder;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 
 import com.squareup.otto.Subscribe;
 
