@@ -1,6 +1,6 @@
 package de.stephanlindauer.criticalmaps.utils;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.io.UnsupportedEncodingException;
 import java.security.MessageDigest;
