@@ -1,7 +1,7 @@
 package de.stephanlindauer.criticalmaps.utils;
 
 import android.os.Environment;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.io.File;
 import java.io.FileInputStream;
