@@ -1,13 +1,27 @@
 package de.stephanlindauer.criticalmaps;
 
-import android.app.Application;
-import android.test.ApplicationTestCase;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
 
-/**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
- */
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
-        super(Application.class);
+import androidx.test.filters.LargeTest;
+import androidx.test.rule.ActivityTestRule;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+
+@RunWith(AndroidJUnit4.class)
+@LargeTest
+public class ApplicationTest {
+
+    @Rule
+    public ActivityTestRule<Main> mActivityRule = new ActivityTestRule<>(Main.class);
+
+    @Test
+    public void verifyAppLaunches() {
+        onView(withId(R.id.drawer_layout)).check(matches(isDisplayed()));
     }
 }
