@@ -1,7 +1,10 @@
 package de.stephanlindauer.criticalmaps;
 
 import android.app.Application;
-import android.support.annotation.NonNull;
+
+import com.squareup.leakcanary.LeakCanary;
+
+import org.jetbrains.annotations.NotNull;
 
 import timber.log.Timber;
 
@@ -12,6 +15,13 @@ public class App extends Application {
     @Override
     public void onCreate() {
         super.onCreate();
+
+        if (LeakCanary.isInAnalyzerProcess(this)) {
+            // This process is dedicated to LeakCanary for heap analysis.
+            // You should not init your app in this process.
+            return;
+        }
+        LeakCanary.install(this);
 
         if (BuildConfig.DEBUG) {
             Timber.plant(new Timber.DebugTree());
@@ -28,7 +38,7 @@ public class App extends Application {
 
     private static class NoOpTree extends Timber.Tree {
         @Override
-        protected void log(int priority, String tag, @NonNull String message, Throwable t) {
+        protected void log(int priority, String tag, @NotNull String message, Throwable t) {
         }
     }
 }
